Update docker-compose.yml and package.json
Update docker-compose.yml to specify the docker-compose version as '1.0' and update the image for the 'postgress' service to use the latest version of Postgres.
Update package.json to use the latest version of '@prisma/client' package.

Add new admin user in seed.ts

Add a new admin user with the username 'admin' in the seed.ts file.